feat: Support custom user commands, loading animations
The new `command` action accepts two parameters, like macros, so that commands
can be identified either by title or (more efficiently) by the cncjs ID assigned
to them.
When a command runs, the button which triggered it (and all other buttons with
the same action) will be put into a loading state, which will end when the command
completes,
whether successfully or unsuccessfully.

In both the web and Stream Deck version, this loading animation is a simple
rotating ring, using the configured progress color. This animation and color
configuration may change in the future

issue #8